requirements.txt
argparse
tensorflow-gpu==2.5.0
# wandb
git+https://github.com/wandb/client.git@summerhack/launch#egg=wandb[launch]